pyproject.toml
[project]
name = "discord-rss-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]
